Changed from doubles to floats

+++ Game/src/game/world/dimensions/Complement.java
@@ -15,7 +15,7 @@ public class Complement extends Rectangle implements Dimension{
 	}
 	
 	@Override
-	public boolean contains(double x, double y){
+	public boolean contains(float x, float y){
 		return dimensionOne.contains(x, y) && !dimensionTwo.contains(x, y);
 	}
