+++ Game/src/game/world/dimensions/Intersection.java
@@ -16,7 +16,7 @@ public class Intersection extends Rectangle implements Dimension{
 	}
 	
 	@Override
-	public boolean contains(double x, double y){
+	public boolean contains(float x, float y){
 		return dimensionOne.contains(x, y) && dimensionTwo.contains(x, y);
 	}
