+++ Game/src/game/world/dimensions/Rectangle.java
@@ -8,12 +8,12 @@ package game.world.dimensions;
  */
 public class Rectangle implements Dimension{
 	
-	private final double x;
-	private final double y;
-	private final double width;
-	private final double height;
+	private final float x;
+	private final float y;
+	private final float width;
+	private final float height;
 	
-	public Rectangle (double x, double y, double width, double height){
+	public Rectangle (float x, float y, float width, float height){
 		this.x = x;
 		this.y = y;
 		this.width = width;
@@ -21,7 +21,7 @@ public class Rectangle implements Dimension{
 	}
 	
 	@Override
-	public boolean contains(double x, double y) {
+	public boolean contains(float x, float y) {
 		if (x < this.x || y < this.y || x > this.width || y > this.height){
 			return false;
 		}
@@ -34,22 +34,22 @@ public class Rectangle implements Dimension{
 	}
 
 	@Override
-	public double getX() {
+	public float getX() {
 		return x;
 	}
 
 	@Override
-	public double getY() {
+	public float getY() {
 		return y;
 	}
 
 	@Override
-	public double getWidth() {
+	public float getWidth() {
 		return width;
 	}
 
 	@Override
-	public double getHeight() {
+	public float getHeight() {
 		return height;
 	}
